First implementation of reflection-based fixG
This commit adds an implementation of our fixpoint combinator based on
reflection, following the observation that the only thing that changes
between all our implementations is the type signature.

Currently, the reflective implementation requires that the user specify
a type signature. It would probably be a good idea to abstract over this
as well.

I'm not particularly happy with how complex the reflective term
definition looks. Maybe there is room for simplifying the definition via
judicious use of quoting/unquoting.