Add Filter example
We introduce the example of filtering streams based on slightly
unrealistic assumptions. This doesn't add much over the Graph example,
but it's (a) simpler and (b) demonstrates the use of fixM-unfold. I'll
also probably end up using this as the introductory example for the
paper.

A variant of this example was present in the codebase earlier.