Add proof that `filter p xs` is a substream of `xs`
The proof serves as an example of how to use fixM-unfold (via
filter-unfold). We give variations for two definitions of the substream
property, one direct and one as an indexed M-type.

A complication arises from the fact that propositional equality of all
fields of two inhabitants of a coinductive record type currently does not
imply propositional equality of the two inhabitants. For the time being,
we add this as an axiom.