Introduce extensional equality for M
There are two problems with propositional equality over `M c s`:

- Proving that two containers are propositionally equal usually
  requires functional extensionality (since the second component of
  a non-indexed container is a function).
- It is currently not provable in Agda that
    inf x == inf y -> x == y
  for `x y : M c s` since we cannot destruct `x` and `y`.

Hence, we would need two axioms to effectively work with M-types.

This commit implements a more laborious but axiom-free alternative: We
define an equality on M-types which abstracts from these details. Users
must then prove that their functions and relations respect this
equality, which is easy (if laborious); this commit adds an example of
how this works in `Filter/`

Todo: Provide an interface that allows users to switch back to
propositional equality if they wish to use the axioms.