Add nicer variant of fixM-unfold using axioms
By using functional extensionality and a sort of extensionality for
coinductive records, we can make fixM-unfold significantly easier to
use since equality of M-types and container functors reduces to simple
propositional (or heterogeneous) equality.

This commit implements an alternative to fixM-unfold which uses the
axioms. On the way, we prove a bunch of lemmas about our own notions of
equality for containers and M-types.