[INCONSISTENT] Require ext only at specific levels
In M.Indexed, we previously required heterogeneous functional
extensionality at all levels. This commit is an experiment trying to
change that, generalising the relevant functions so that they require
extensionality only at specific levels. However, as it turns out, this
change results in pretty ridiculous complexity, so the experiment
remains unfinished. I record it in this commit in case we want to make
another attempt later.