Separate library and examples
Now that I have a decently clear idea of what's going into the library,
we can set up a less chaotic directory structure. To that end, this
commit

- moves the library into a new `src` directory;
- moves all library modules to `Coinduction.WellFounded.*` to minimize the risk
  of name collisions;
- creates a `.agda-lib` file for the library;
- moves all the examples to a new `examples` directory;
- adds info on how to build the examples.

The generic approach (formerly in `Generic.agda`) is out for the time
being because I can't be arsed to port the fixpoint lemma. Agda's
reflection is just too horrible in practice.