Add runLengths example
`runLengths : Stream A -> Stream A` takes a stream and replaces each run
of decreasing elements with its length. It's a pretty good example for
our technique because you would naturally appeal to well-founded
recursion in a productivity argument, and fixM allows us to reflect this
directly in the code. Besides, the example is much less contrived than
filtering.