Add property to runs example
The runs example now contains a proof of a simple property about `runs`,
namely that the output stream of lists contains only elements from the
input stream. This holds trivially by parametricity, but the proof is
interesting because it requires the indexed fixM.

On the side, we move some basic stuff from `M.agda` to the new
`Base.agda`. We also change the order of arguments for `runs'` to have
the stream in front. This lets us generalise _<<_, which doesn't
currently help much but may be useful in the future.