Move internal definitions to *.Internal files
We used to tuck internal definitions away into Internal_i modules (i a
number) in the same file as the 'external' definitions. This has the
disadvantage that opening the enclosing module unqualified brings
Internal_i into scope as well, which easily leads to name conflicts if
multiple modules follow this model.

To avoid this problem, the present commit switches to a style that is
popular among Haskellers: Everything is defined in internal modules --
say, `Thing.Internal` --, and `Thing` only re-exports the public
definitions.

Another alternative would be to simply mark internal definitions as
`private`. But this is problematic because library clients who need
access to the definitions due to unforeseen circumstances can't get it,
which is even more inconvenient in Agda than in Haskell because proofs
frequently need access to implementation details.

The present solution, therefore, sits in a sweet spot: It clearly
demarcates the boundary between public and private definitions (with
only the former covered by the versioning contract) while
inconveniencing users as little as possible.