Update examples for Agda 2.5.3/stdlib-0.14
Version 0.14 of the standard library changed some definitions concerning
lists and natural numbers which we use for the examples. In particular,
well-founded induction now uses _<_ rather than _<'_, which allows us to
drop a good bunch of code in Runs.Nat.

The update to Agda 2.5.3 necessitates only small changes. In particular,
it appears that size inference has become a little less generous and the
syntax for record type declarations has changed slightly.